+++ src/gui/ServerGUI.java
@@ -1,4 +1,4 @@
-package pkg445group;
+package gui;
 
 import java.awt.BorderLayout;
 import java.awt.event.ActionEvent;
